"Refactor exam entity"

+++ SpringBootDemo200301/src/main/java/com/hqyj/SpringBootDemo/modules/exam/entity/Exam.java
@@ -23,7 +23,7 @@ public class Exam {
 	private int paperId;
 	// 参考得分
 	private String referenceScore;
-	private int score;
+	private Double score;
 	private int spendTime;
 	private LocalDateTime examDate;
 
@@ -59,11 +59,11 @@ public class Exam {
 		this.referenceScore = referenceScore;
 	}
 
-	public int getScore() {
+	public Double getScore() {
 		return score;
 	}
 
-	public void setScore(int score) {
+	public void setScore(Double score) {
 		this.score = score;
 	}
 
@@ -82,4 +82,5 @@ public class Exam {
 	public void setExamDate(LocalDateTime examDate) {
 		this.examDate = examDate;
 	}
+
 }
